fixed bug #4437: Rating wikis as anonymous user
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@20234 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Services/Rating/classes/class.ilRating.php
@@ -29,6 +29,11 @@ class ilRating
 	{
 		global $ilDB;
 
+		if ($a_user_id == ANONYMOUS_USER_ID)
+		{
+			return;
+		}
+
 		$ilDB->manipulate("DELETE FROM il_rating WHERE ".
 			"user_id = ".$ilDB->quote($a_user_id, "integer")." AND ".
 			"obj_id = ".$ilDB->quote((int) $a_obj_id, "integer")." AND ".
